+++ gui_widgets/D3Frame.py
@@ -1,11 +1,14 @@
 from tkinter import *
 from gui_widgets.FrameBaseClass import FrameBaseClass
+from gui_widgets.SpectralMap3d import Render3DGraph
 
 # # TODO LOGIKA PRI ZMENE CONNECTION OBRAZKA
 
 class D3Frame(FrameBaseClass):
-    def __init__(self):
+    def __init__(self, motorFrame):
         super().__init__()
+
+        self.motorFrame = motorFrame
 
         # Setting color of frame
         self.configure(bg=self.FRAME_COLOR)
@@ -19,7 +22,7 @@ class D3Frame(FrameBaseClass):
 
         # Buttons
         self.showButton = self.initializeButton(self.BUTTON_SIZE_HEIGHT, self.BUTTON_SIZE_WIDTH, "Show")
-        self.showButton.configure(command=lambda: self.FUNCTION_TODO("ARGUMENT"))
+        self.showButton.configure(command=lambda: self.show3Dgraph())
 
         # Images
         self.scanImage = Label(self, image=self.OFF_IMAGE, bg=self.FRAME_COLOR)
@@ -40,3 +43,9 @@ class D3Frame(FrameBaseClass):
         self.showButton.grid(row=2, column=1, padx=(0, 10), pady=(10, 0))
         self.scanLabel.grid(sticky=W, row=3, column=0, padx=(0, 10), pady=(10, 10))
         self.scanImage.grid(row=3, column=0, padx=(30, 10), pady=(10, 10))
+
+    def show3Dgraph(self):
+        dataFor3D = self.motorFrame.motorController.dataContainer
+        render = Render3DGraph(dataFor3D)
+        print("data of controller", dataFor3D)
+        render.renderHeightMap(550)
